# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SRI - Release Go Live Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List release test scenarios
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: prioridade
          in: query
          required: false
          schema:
            type: string
        - name: site
          in: query
          required: false
          schema:
            type: string
        - name: sistema
          in: query
          required: false
          schema:
            type: string
        - name: statusCenario
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: createScenario
      tags: [scenarios]
      summary: Create a scenario (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "201":
          description: Created scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get a scenario by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateScenario
      tags: [scenarios]
      summary: Update a scenario (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioUpdate"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteScenario
      tags: [scenarios]
      summary: Delete a scenario (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /scenarios/{id}/status:
    patch:
      operationId: updateScenarioStatus
      tags: [scenarios]
      summary: Update only the status of a scenario
      description: Requires the cenarios:alterar_status permission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioStatusUpdate"
      responses:
        "200":
          description: Updated scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an evidence file from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir.
      responses:
        "200":
          description: Object content streamed with correct Content-Type.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Scenario:
      type: object
      required: [id, idTeste]
      properties:
        id:
          type: integer
        idTeste:
          type: string
        chaveamento:
          type: ["string", "null"]
        sequencia:
          type: ["string", "null"]
        blocoExecucao:
          type: ["string", "null"]
        fisicoSistemico:
          type: ["string", "null"]
        prioridade:
          type: ["string", "null"]
        cenario:
          type: ["string", "null"]
        dependenciaCenarioExterno:
          type: ["string", "null"]
        quemExecuta:
          type: ["string", "null"]
        baselineCustomizado:
          type: ["string", "null"]
        facilitador:
          type: ["string", "null"]
        keyUser:
          type: ["string", "null"]
        superUser:
          type: ["string", "null"]
        endUser:
          type: ["string", "null"]
        macroProcesso:
          type: ["string", "null"]
        sequenciaPassoAPasso:
          type: ["string", "null"]
        evidenciasObrigatorias:
          type: ["string", "null"]
        quemDefineMassa:
          type: ["string", "null"]
        massaDados:
          type: ["string", "null"]
        celula:
          type: ["string", "null"]
        agrupamento:
          type: ["string", "null"]
        tipoCenario:
          type: ["string", "null"]
        liberacao:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        sistema:
          type: ["string", "null"]
        site:
          type: ["string", "null"]
        statusCenario:
          type: ["string", "null"]
        idDefeitoJira:
          type: ["string", "null"]
        idCenarioJira:
          type: ["string", "null"]
        statusCheckPoint:
          type: ["string", "null"]
        diretorio:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
    ScenarioInput:
      type: object
      required: [idTeste]
      properties:
        idTeste:
          type: string
          minLength: 1
        chaveamento:
          type: string
        sequencia:
          type: string
        blocoExecucao:
          type: string
        fisicoSistemico:
          type: string
        prioridade:
          type: string
        cenario:
          type: string
        dependenciaCenarioExterno:
          type: string
        quemExecuta:
          type: string
        baselineCustomizado:
          type: string
        facilitador:
          type: string
        keyUser:
          type: string
        superUser:
          type: string
        endUser:
          type: string
        macroProcesso:
          type: string
        sequenciaPassoAPasso:
          type: string
        evidenciasObrigatorias:
          type: string
        quemDefineMassa:
          type: string
        massaDados:
          type: string
        celula:
          type: string
        agrupamento:
          type: string
        tipoCenario:
          type: string
        liberacao:
          type: string
        observacoes:
          type: string
        sistema:
          type: string
        site:
          type: string
        statusCenario:
          type: string
        idDefeitoJira:
          type: string
        idCenarioJira:
          type: string
        statusCheckPoint:
          type: string
        diretorio:
          type: string
    ScenarioUpdate:
      type: object
      properties:
        idTeste:
          type: string
        chaveamento:
          type: string
        sequencia:
          type: string
        blocoExecucao:
          type: string
        fisicoSistemico:
          type: string
        prioridade:
          type: string
        cenario:
          type: string
        dependenciaCenarioExterno:
          type: string
        quemExecuta:
          type: string
        baselineCustomizado:
          type: string
        facilitador:
          type: string
        keyUser:
          type: string
        superUser:
          type: string
        endUser:
          type: string
        macroProcesso:
          type: string
        sequenciaPassoAPasso:
          type: string
        evidenciasObrigatorias:
          type: string
        quemDefineMassa:
          type: string
        massaDados:
          type: string
        celula:
          type: string
        agrupamento:
          type: string
        tipoCenario:
          type: string
        liberacao:
          type: string
        observacoes:
          type: string
        sistema:
          type: string
        site:
          type: string
        statusCenario:
          type: string
        idDefeitoJira:
          type: string
        idCenarioJira:
          type: string
        statusCheckPoint:
          type: string
        diretorio:
          type: string
    ScenarioStatusUpdate:
      type: object
      required: [statusCenario]
      properties:
        statusCenario:
          type: string
          minLength: 1
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"