# lib/api-zod/src/generated/types/scenarioStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SRI - Release Go Live Control API
 * OpenAPI spec version: 0.1.0
 */

export interface ScenarioStatusUpdate {
  /** @minLength 1 */
  statusCenario: string;
}